answers for assignment
1.What is a .gitignore file, and what should a Java project's include in one?

Ans: The .gitignore file will be in the repo's root folder, and we'll add the file extensions we want to exclude from the repo to it. We must include the *.jar and *.class files.

2.What does Git's fast-forward mean?

Ans: Merging changes from another branch into your local branch without causing conflicts is a fast-forward.
3.Why do you need multiple branches?

Ans: To avoid frequent conflicts between members, everyone should be responsible for their own changes and keep in close contact with the development branch's master.

4. How many remote Git repositories do you generally have in a project?

Ans: There is no limit, and you can make multiple repos depending on your needs.